backend/requirements.txt
# Core Flask packages
flask>=2.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.0
python-dotenv>=1.0.0
werkzeug>=2.0.0

# Database
sqlalchemy>=2.0.0

# Security
bcrypt>=4.0.0

# HTTP and API
requests>=2.31.0

# AI and ML packages
google-generativeai>=0.3.0

# Basic data science (lightweight versions)
numpy>=1.24.0
pandas>=2.0.0

# Optional ML packages (install only if needed)
# transformers>=4.35.0
# torch>=2.2.0
# scikit-learn>=1.3.0
# nltk>=3.8.0
# spacy>=3.7.0

# Optional multimedia packages
# speech-recognition>=3.10.0
# pydub>=0.25.0
# Pillow>=10.0.0

# Optional cloud packages
# google-cloud-speech>=2.21.0
# google-cloud-drive>=3.0.0

# Optional task queue
# celery>=5.3.0
# redis>=5.0.0

# Optional visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0
# plotly>=5.17.0

# Development tools
# jupyter>=1.0.0
# ipython>=8.16.0

# Production server
gunicorn>=21.0.0
werkzeug==2.3.7
